优化了一下缓存清理方案

+++ src/index.ts
@@ -10,12 +10,12 @@ export interface NodeDiskCacheOptions {
     cacheDir?: string;
 
     /**
-     * 缓存容量上限(byte)，默认为缓存目录可用容量 * volumeUpLimitRate
+     * 缓存容量上限(byte)，默认为0，没有上限
      */
     volumeUpLimit?: number;
 
     /**
-     * 缓存占所在目录可用容量的上限比例，默认为0.9。如果设置了volumeUpLimit则会使该属性失效
+     * 动态监测缓存目录剩余容量，当已用容量占总容量超过指定比例后执行清理操作。范围0-1，默认0，没有上限。如果设置了volumeUpLimit则会使该属性失效
      */
     volumeUpLimitRate?: number;
 
@@ -38,14 +38,17 @@ export default class NodeDiskCache {
     //缓存数据存放目录
     private readonly _cacheDir: string;
 
+    //缓存目录列表，防止某一缓存目录被重复使用
+    private static readonly _cacheDirList = new Set<string>();
+
     //缓存超时
     private readonly _timeout: number;
 
     //是否获取缓存时重置timeout
     private readonly _refreshTimeoutWhenGet: boolean;
 
-    //缓存容量上限
-    private readonly _volumeUpLimit: number;
+    //清理缓存计时器
+    private readonly _cleanerTimer: NodeJS.Timer;
 
     //当前容量
     private _currentVolume = 0;
@@ -56,17 +59,52 @@ export default class NodeDiskCache {
     constructor(options: NodeDiskCacheOptions = {}) {
         //缓存目录
         this._cacheDir = options.cacheDir || path.join(os.tmpdir(), `NodeDiskCache_${Math.trunc(Math.random() * 10000)}`);
+        if (NodeDiskCache._cacheDirList.has(this._cacheDir))
+            throw new Error(`缓存目录已被占用：${this._cacheDir}`);
         fs.emptyDirSync(this._cacheDir);
+        NodeDiskCache._cacheDirList.add(this._cacheDir);
 
-        //缓存容量
-        if (options.volumeUpLimit as number > 0)
-            this._volumeUpLimit = options.volumeUpLimit as number;
-        else
-            this._volumeUpLimit = diskusage.checkSync(this._cacheDir).available * Math.min(Math.max(options.volumeUpLimitRate || 0.9, 0), 1);
+        //清理缓存
+        if (options.volumeUpLimit as number > 0) {
+            const upLimit = options.volumeUpLimit as number;
+            this._cleanerTimer = setInterval(() => {
+                if (this._currentVolume > upLimit)
+                    this._cleanCache(upLimit * 0.9);
+            }, 5000);
+        } else if (options.volumeUpLimitRate as number > 0) {
+            const downLimitRate = 1 - Math.min(options.volumeUpLimitRate as number, 1);
+            this._cleanerTimer = setInterval(async () => {
+                try {
+                    const usage = await diskusage.check(this._cacheDir);
+                    if (usage.available / usage.total < downLimitRate)
+                        this._cleanCache(this._currentVolume - usage.total * 0.1);
+                } catch (err) {
+                    console.error('获取缓存目录容量信息异常：', err);
+                }
+            }, 5000);
+        }
 
         //缓存超时
         this._timeout = options.timeout as number > 0 ? options.timeout as number : 0;
         this._refreshTimeoutWhenGet = !!options.refreshTimeoutWhenGet;
+    }
+
+    /**
+     * 清理缓存
+     * @param downTo 将缓存大小下降到指定数值之下
+     */
+    private _cleanCache(downTo: number): void {
+        (async () => {
+            for (const [key, value] of this._cacheTable) {
+                if (this._currentVolume > downTo) {
+                    this._cacheTable.delete(key);
+                    clearTimeout(value.timer as any);
+                    await fs.remove(value.name);
+                    this._currentVolume -= value.size;
+                } else
+                    break;
+            }
+        })().catch(err => console.error('清除缓存异常：', err));
     }
 
     /**
@@ -75,8 +113,8 @@ export default class NodeDiskCache {
     async set(key: string, value: Buffer | NodeJS.ReadableStream): Promise<void> {
         const cache = this._cacheTable.get(key) || { name: path.join(this._cacheDir, (this._fileNameIndex++).toString()), size: 0, timer: undefined };
 
-        if (this._timeout > 0)
-            clearTimeout(cache.timer as any);
+        if (this._timeout > 0 && cache.timer !== undefined)
+            clearTimeout(cache.timer);
 
         //保存缓存
         if (Buffer.isBuffer(value)) {
@@ -95,23 +133,6 @@ export default class NodeDiskCache {
             cache.size = status.size;
         }
 
-        //判断缓存是否已经超过了容量限制
-        if (this._currentVolume > this._volumeUpLimit) {
-            const downTo = this._volumeUpLimit * 0.9;   //将缓存下降到
-
-            for (const [ckey, value] of this._cacheTable) {
-                if (key === ckey)
-                    continue;
-                else if (this._currentVolume > downTo) {
-                    this._cacheTable.delete(ckey);
-                    clearTimeout(value.timer as any);
-                    await fs.remove(value.name);
-                    this._currentVolume -= value.size;
-                } else
-                    break;
-            }
-        }
-
         if (this._timeout > 0) {
             cache.timer = setTimeout(() => {
                 this._cacheTable.delete(key);
@@ -124,6 +145,7 @@ export default class NodeDiskCache {
             }, this._timeout);
         }
 
+        this._cacheTable.delete(key);   //刷新缓存在列表中的排位
         this._cacheTable.set(key, cache);
     }
 
@@ -148,8 +170,7 @@ export default class NodeDiskCache {
             }
 
             return await fs.readFile(cache.name);
-        }
-        else
+        } else
             return cache;
     }
 
@@ -210,5 +231,13 @@ export default class NodeDiskCache {
             clearTimeout(value.timer as any);
             this._currentVolume -= value.size;
         }
+    }
+
+    /**
+     * 销毁缓存
+     */
+    async destroy(): Promise<void> {
+        await this.empty();
+        clearInterval(this._cleanerTimer);
     }
 }
